refactor(test): unify Hive and spec-test fork-choice runner
